add combine method to reducers
