+++ src/actions/index.js
@@ -1,0 +1,9 @@
+export const createBook = (book) => {
+  type = 'CREATE_BOOK',
+  book
+}
+
+export const removeBook = (book) => {
+  type = 'REMOVE_BOOK',
+  book
+}
