+++ src/components/BooksList.js
@@ -1,6 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import Book from './Book';
+import { connect } from 'react-redux';
 
 function BooksList(props) {
   const { books } = props;
@@ -22,4 +23,10 @@ BooksList.propTypes = {
   books: PropTypes.isRequired,
 };
 
-export default BooksList;
+function mapStateToProps(state) {
+  return {
+    books: state.books,
+  }
+}
+
+export default connect(mapStateToProps)(BooksList);
