+++ src/reducers/index.js
@@ -1,0 +1,4 @@
+import books from './books';
+import { combineReducers } from 'redux';
+
+export default combineReducers({ books });
